refactor: extract test helpers into feature-gated `test-utils` module
